+++ CHANGES.md
@@ -4,6 +4,6 @@
 * Removed `@expectNothing`
 * Now failes if examples produces outcomes that are not handled
 * `--filter` now only works on named examples
-* Added the `--named-only` option
+* Added the `--flagged-only` option
 * Added the `--format` option and the json format
 * Added the `--ignore-unknown-annotations` option